Fixed the rendering on the block details and the nav bar
