+++ src/malcolm/reducer/malcolmReducer.js
@@ -14,7 +14,9 @@ import {
   MalcolmReturn,
 } from '../malcolm.types';
 import { AlarmStates } from '../../malcolmWidgets/attributeDetails/attributeAlarm/attributeAlarm.component';
-import NavigationReducer from './navigation.reducer';
+import NavigationReducer, {
+  processNavigationLists,
+} from './navigation.reducer';
 import AttributeReducer from './attribute.reducer';
 
 const initialMalcolmState = {
@@ -63,6 +65,7 @@ function registerNewBlock(state, action) {
 
   if (!Object.prototype.hasOwnProperty.call(blocks, action.payload.blockName)) {
     blocks[action.payload.blockName] = {
+      attributes: [],
       name: action.payload.blockName,
       loading: true,
       children: [],
@@ -83,6 +86,7 @@ function registerNewBlock(state, action) {
 
 function updateBlock(state, payload) {
   const blocks = { ...state.blocks };
+  let { navigation } = state;
 
   if (payload.delta) {
     const blockName = state.messagesInFlight.find(m => m.id === payload.id)
@@ -98,14 +102,26 @@ function updateBlock(state, payload) {
           loading: true,
           children: [],
         })),
-        children: ['none', ...payload.fields],
+        children: [...payload.fields],
       };
+    }
+
+    if (
+      state.navigation
+        .map(nav => nav.path)
+        .findIndex(path => path === blockName) > -1
+    ) {
+      navigation = processNavigationLists(
+        state.navigation.map(nav => nav.path),
+        blocks
+      );
     }
   }
 
   return {
     ...state,
     blocks,
+    navigation,
   };
 }
